refactor: 폴더 삭제 후 파일명 수정
